pyproject.toml
[project]
name = "sarthiai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "cloudinary>=1.44.2",
    "dj-database-url>=3.1.2",
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "django-rest-framework>=0.1.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.3.0",
    "pillow>=12.2.0",
    "psycopg>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "whitenoise>=6.12.0",
]
